pyproject.toml
[project]
name = "box-agent"
version = "0.8.24"
description = "AI agent framework with Jupyter sandbox, data analysis, MCP tools, ACP protocol, multi-provider LLM, and standalone runtime packaging"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    {name = "Box Agent Team"}
]
license = "MIT"
keywords = ["agent", "ai", "llm", "sandbox", "mcp", "acp", "tool-calling", "safety", "jupyter", "data-analysis", "electron", "standalone-runtime"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "pydantic>=2.0.0",
    "pyyaml>=6.0.0",
    "httpx>=0.27.0",
    "mcp>=1.0.0",
    "requests>=2.31.0",
    "tiktoken>=0.5.0",
    "prompt-toolkit>=3.0.0",
    "anthropic>=0.39.0",
    "openai>=1.57.4",
    "agent-client-protocol>=0.6.0",
    "jupyter-client>=8.6.0",
    "ipykernel>=6.29.0",
]

[project.scripts]
box-agent = "box_agent.cli:main"
box-agent-acp = "box_agent.acp.server:main"

[project.urls]
Homepage = "https://github.com/Raccoon-Office/Box-Agent"
Repository = "https://github.com/Raccoon-Office/Box-Agent"
Documentation = "https://github.com/Raccoon-Office/Box-Agent#readme"
Issues = "https://github.com/Raccoon-Office/Box-Agent/issues"
Changelog = "https://github.com/Raccoon-Office/Box-Agent/releases"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]
runtime = [
    "pandas",
    "numpy",
    "matplotlib",
    "seaborn",
    "openpyxl",
    "xlrd",
    "scikit-learn",
    "python-docx",
    "pypdf",
    "pdfplumber",
    "reportlab",
    "python-pptx",
    "pip",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["box_agent*"]
exclude = ["tests*"]

# Include skills directory (git submodule) as package data
# After moving submodule to box_agent/skills
[tool.setuptools.package-data]
box_agent = ["skills/**/*", "config/*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
cache_dir = "workspace/.pytest_cache"
asyncio_mode = "auto"

[tool.pylint.messages_control]
# Disable warnings for Tool.execute method signature differences
# We intentionally use explicit parameters in subclasses for better type hints
disable = [
    "arguments-differ",  # Allow subclasses to have different parameter signatures
]

[dependency-groups]
dev = [
    "pytest-asyncio>=1.2.0",
    "pytest-cov>=7.0.0",
    "pytest-xdist>=3.8.0",
    "pyinstaller>=6.0.0",
]
